Issue #5, Moved the mock graph data service to a separate file, added documentation
